refactor to index.js
