+++ src/Screens/DashboardScreen/style.js
@@ -1,0 +1,12 @@
+import { StyleSheet } from 'react-native'
+
+const styles = StyleSheet.create({
+    view: {
+        flex: 1,
+        alignItems: 'center',
+        justifyContent: 'center',
+        backgroundColor: 'white',
+    },
+});
+
+export default styles;
