+++ src/components/screens/AuthenticationScreen/AuthenticationTextInput/index.js
